// product/frontend/DNAustria/src/app/api/model/organizationDto.ts
/**
 * DNAustria.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AddressDto } from './addressDto';


export interface OrganizationDto { 
    id: number;
    name: string;
    adress: AddressDto;
}