// product/frontend/DNAustria/src/app/api/model/updateEventDto.ts
/**
 * DNAustria.Api | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface UpdateEventDto { 
    name: string;
    description: string;
    link: string;
    startDate: string;
    endDate: string;
    classification: number;
    status: number;
    hasFees: boolean;
    isOnline: boolean;
    organization: number | null;
    programName?: string;
    format?: string;
    schoolBookable: boolean;
    ageMinimum: number;
    ageMaximum: number;
    location: number | null;
    contact: number | null;
    targetAudiences?: Array<number> | null;
    topics?: Array<number> | null;
}